+++ src/users/viewsets/user/view.py
@@ -6,7 +6,7 @@ from rest_framework.response import Response
 from core.services.email import EmailService
 from core.utils.file import replace_keys
 from users.models import User
-from users.viewsets.user.permissions import IsAllowed
+from core.permissions.is_allowed import IsAllowed
 from users.viewsets.user.serializer import UserSerializer
 from users.viewsets.user.service import RegisterUserService
 from django.conf import settings
